Feat/devise pundit

+++ app/controllers/books_controller.rb
@@ -1,14 +1,16 @@
 # frozen_string_literal: true
 
 class BooksController < ApplicationController
+  before_action :authenticate_user!
   before_action :find_single_book, only: %i[show edit update destroy]
 
   def index
-    @books = Book.all
+    @books = Book.order(:name).page(params[:page]).per(10)
+    @total_books = Book.count
   end
 
   def show
-    @reviews = Review.where(book_id: params[:id])
+    @reviews = Review.where(book_id: params[:id]).page(params[:page]).per(10)
   end
 
   def new
@@ -16,7 +18,8 @@ class BooksController < ApplicationController
   end
 
   def create
-    @book = Book.new(book_params)
+    @user = User.find(current_user.id)
+    @book = @user.books.build(book_params)
     @book.save!
     redirect_to @book
   end
@@ -24,12 +27,14 @@ class BooksController < ApplicationController
   def edit; end
 
   def update
+    authorize @book, policy_class: BookPolicy
     @book.update!(book_params)
     flash[:success] = 'Book update completed'
     redirect_to @book
   end
 
   def destroy
+    authorize @book, policy_class: BookPolicy
     flash[:alert] = 'Failed to delete the book' unless @book.destroy
     redirect_to root_path
   end
